+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 next
 =====
 
+- `IniModifier::save()`: support of new flag `FORMAT_SPACE_AROUND_EQUAL` to add
+  spaces around the equal sign
 - new method `Utils::arrayToIniString()`
 - Fix an issue when given `''` as a section name to methods of `IniModifier`.
 - The library is compatible with PHP 8.2 and 8.3.